# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ad Seller System API",
    "description": "IAB OpenDirect 2.1 compliant seller agent for programmatic advertising. Supports product discovery, tiered pricing, proposal evaluation, multi-round negotiation, deal execution, order management, and change requests.",
    "contact": {
      "name": "IAB Tech Lab",
      "url": "https://iabtechlab.com/"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/media-kit/packages": {
      "get": {
        "tags": [
          "Media Kit"
        ],
        "summary": "List Media Kit Packages",
        "description": "List packages with public view (price ranges, no exact pricing).\n\nAccepts the same audience-filter triple as `GET /packages` so public\ndiscovery callers can narrow by audience type without authenticating.",
        "operationId": "list_media_kit_packages_media_kit_packages_get",
        "parameters": [
          {
            "name": "layer",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Layer"
            }
          },
          {
            "name": "featured_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Featured Only"
            }
          },
          {
            "name": "audience_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audience Type"
            }
          },
          {
            "name": "audience_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audience Id"
            }
          },
          {
            "name": "audience_taxonomy_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audience Taxonomy Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/media-kit/packages/{package_id}": {
      "get": {
        "tags": [
          "Media Kit"
        ],
        "summary": "Get Media Kit Package",
        "description": "Get a single package with public view.",
        "operationId": "get_media_kit_package_media_kit_packages__package_id__get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Package Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/media-kit/search": {
      "post": {
        "tags": [
          "Media Kit"
        ],
        "summary": "Search Media Kit",
        "description": "Search packages by keyword. Authenticated buyers get richer results.\n\nPer proposal \u00a75.7, the scoring corpus now includes\n`audience_capabilities.standard_segment_ids` +\n`audience_capabilities.contextual_segment_ids` alongside keywords/tags\n-- a query mentioning a known IAB segment ID ranks packages that\ndeclare it higher than packages that don't.\n\nThe optional `audience_filter` body field restricts results to packages\nthat match its type/id/version triple, parallel to `GET /packages`.",
        "operationId": "search_media_kit_media_kit_search_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaKitSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/packages": {
      "get": {
        "tags": [
          "Packages"
        ],
        "summary": "List Packages",
        "description": "List packages with tier-gated view.\n\nAudience filter (proposal \u00a75.7):\n\n- `audience_type`: one of `standard` | `contextual` | `agentic`.\n- `audience_id`: taxonomy ID for standard/contextual; URI for agentic.\n  Requires `audience_type` to disambiguate which capability list to\n  search.\n- `audience_taxonomy_version`: optional version constraint; when unset\n  the seller's lock-file version is authoritative.\n\nEmpty results return `[]`, not 404 -- matches the existing behavior for\nlayer/featured filters.",
        "operationId": "list_packages_packages_get",
        "parameters": [
          {
            "name": "buyer_tier",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "public",
              "title": "Buyer Tier"
            }
          },
          {
            "name": "agency_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agency Id"
            }
          },
          {
            "name": "advertiser_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Advertiser Id"
            }
          },
          {
            "name": "layer",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Layer"
            }
          },
          {
            "name": "audience_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audience Type"
            }
          },
          {
            "name": "audience_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audience Id"
            }
          },
          {
            "name": "audience_taxonomy_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audience Taxonomy Version"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Packages"
        ],
        "summary": "Create Package",
        "description": "Create a curated package (Layer 2).",
        "operationId": "create_package_packages_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/packages/{package_id}": {
      "get": {
        "tags": [
          "Packages"
        ],
        "summary": "Get Package",
        "description": "Get a single package with tier-gated view.",
        "operationId": "get_package_packages__package_id__get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Package Id"
            }
          },
          {
            "name": "buyer_tier",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "public",
              "title": "Buyer Tier"
            }
          },
          {
            "name": "agency_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agency Id"
            }
          },
          {
            "name": "advertiser_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Advertiser Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Packages"
        ],
        "summary": "Update Package",
        "description": "Update an existing package.",
        "operationId": "update_package_packages__package_id__put",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Package Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Updates"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Packages"
        ],
        "summary": "Delete Package",
        "description": "Archive a package (soft delete).",
        "operationId": "delete_package_packages__package_id__delete",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Package Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/packages/assemble": {
      "post": {
        "tags": [
          "Packages"
        ],
        "summary": "Assemble Package",
        "description": "Assemble a dynamic package (Layer 3) from product IDs.",
        "operationId": "assemble_package_packages_assemble_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DynamicPackageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DynamicPackageRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "product_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Product Ids"
          }
        },
        "type": "object",
        "required": [
          "name",
          "product_ids"
        ],
        "title": "DynamicPackageRequest",
        "description": "Request to assemble a dynamic package from product IDs."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaKitSearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "buyer_tier": {
            "type": "string",
            "title": "Buyer Tier",
            "default": "public"
          },
          "agency_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agency Id"
          },
          "advertiser_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advertiser Id"
          },
          "audience_filter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AudienceFilterModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "MediaKitSearchRequest",
        "description": "Request to search packages."
      },
...
    }
  },
...
}